+++ db/migrate/20210323005947_create_pp_criterias_ponderations.rb
@@ -3,7 +3,7 @@ class CreatePpCriteriasPonderations < ActiveRecord::Migration[5.2]
     create_table :pp_criterias_ponderations do |t|
       t.belongs_to :pp_criteria
       t.belongs_to :user
-      t.integer :value
+      t.float :value
     end
   end
 end
